[UT] Move PseudoCluster based scan limit test out of PartitionPruneRuleTest
PseudoCluster.shutdown() only deletes the run dir, so the pseudo FE/BE daemon
threads keep running for the rest of the JVM's life and keep calling OlapTable
methods. Surefire forks one JVM per test class, so those threads outlive
testOlapPartitionScanLimit and run concurrently with the JMockit tests in the
same class, whose @Mocked OlapTable redefines that class JVM-wide. The daemon
calls then mutate JMockit's global record/replay state while the Expectations
block is still recording, which fails with ConcurrentModificationException.

Move the test into its own class so the JMockit tests run in a JVM without a
live cluster.
