+++ fe/fe-core/src/test/java/com/starrocks/sql/optimizer/rule/transformation/PartitionPruneRuleTest.java
@@ -29,7 +29,6 @@ import com.starrocks.catalog.PartitionType;
 import com.starrocks.catalog.RangePartitionInfo;
 import com.starrocks.common.AnalysisException;
 import com.starrocks.common.FeConstants;
-import com.starrocks.pseudocluster.PseudoCluster;
 import com.starrocks.sql.ast.PartitionValue;
 import com.starrocks.sql.ast.expression.BinaryType;
 import com.starrocks.sql.ast.expression.DateLiteral;
@@ -50,11 +49,8 @@ import com.starrocks.type.PrimitiveType;
 import com.starrocks.type.StringType;
 import mockit.Expectations;
 import mockit.Mocked;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-import java.sql.Connection;
-import java.sql.Statement;
 import java.time.LocalDateTime;
 import java.util.HashMap;
 import java.util.List;
@@ -459,60 +455,4 @@ public class PartitionPruneRuleTest {
         long actual = selectPartitionIds.get(0);
         assertEquals(10001L, actual);
     }
-
-    @Test
-    public void testOlapPartitionScanLimit() throws Exception {
-        PseudoCluster.getOrCreateWithRandomPort(true, 1);
-        Connection connection = PseudoCluster.getInstance().getQueryConnection();
-        Statement stmt = connection.createStatement();
-        try {
-            stmt.execute("create database olap_partition_scan_limit_test_db");
-            stmt.execute("use olap_partition_scan_limit_test_db");
-            stmt.execute("CREATE TABLE olap_partition_scan_limit_test_table " +
-                    "(`a` varchar(65533),`b` varchar(65533),`ds` date) ENGINE=OLAP " +
-                    "DUPLICATE KEY(`a`) PARTITION BY RANGE(`ds`)" +
-                    "(START (\"2024-09-20\") END (\"2024-09-27\") EVERY (INTERVAL 1 DAY))" +
-                    "DISTRIBUTED BY HASH(`a`)" +
-                    "PROPERTIES (\"replication_num\" = \"1\")");
-            stmt.execute("insert into olap_partition_scan_limit_test_table(a,b,ds) " +
-                    "values('1','a','2024-09-20'),('2','a','2024-09-21')," +
-                    "('3','a','2024-09-22'),('4','a','2024-09-23'),('5','a','2024-09-24')," +
-                    "('6','a','2024-09-25'),('7','a','2024-09-26')");
-            //check default value 0
-            stmt.execute("select count(*) from olap_partition_scan_limit_test_table where ds>='2024-09-22';");
-            if (stmt.getResultSet().next()) {
-                int count = stmt.getResultSet().getInt(1);
-                Assertions.assertEquals(count, 5);
-            }
-            //check set value -1
-            stmt.execute("set scan_olap_partition_num_limit=-1;");
-            stmt.execute("select count(*) from olap_partition_scan_limit_test_table where ds>='2024-09-22';");
-            if (stmt.getResultSet().next()) {
-                int count = stmt.getResultSet().getInt(1);
-                Assertions.assertEquals(count, 5);
-            }
-            //check set value 3
-            stmt.execute("set scan_olap_partition_num_limit=3;");
-            try {
-                stmt.execute("select count(*) from olap_partition_scan_limit_test_table where ds>='2024-09-22';");
-            } catch (Exception e) {
-                String exp = "Exceeded the limit of number of olap table partitions to be scanned. Number of partitions " +
-                        "allowed: 3, number of partitions to be scanned: 5. Please adjust the SQL or " +
-                        "change the limit by set variable scan_olap_partition_num_limit.";
-                Assertions.assertTrue(e.getMessage().contains(exp));
-            }
-            //check set invalid value abc
-            try {
-                stmt.execute("set scan_olap_partition_num_limit=abc;");
-            } catch (Exception e) {
-                String exp = "Incorrect argument type to variable 'scan_olap_partition_num_limit'";
-                Assertions.assertTrue(e.getMessage().contains(exp));
-            }
-        } finally {
-            stmt.close();
-            connection.close();
-            PseudoCluster.getInstance().shutdown(true);
-        }
-    }
-
 }
